petit printf + debut de ls

+++ printfast/srcs/get_s.c
@@ -1,5 +1,5 @@
 
-#include "printfast.h"
+#include "../includes/printfast.h"
 
 char	*get_s(va_list params, t_att option)
 {
@@ -9,7 +9,7 @@ char	*get_s(va_list params, t_att option)
 
 	str = va_arg(params, char*);
 	len = ft_strlen(str);
-	if (option.width == -1 || option.width >= ft_strlen(str))
+	if (option.width == -1 || option.width >= (int)ft_strlen(str))
 		ret = ft_strdup(str);
 	else
 		ret = ft_strndup(str, option.width);
